requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.1
pydantic-settings==2.1.0
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2
requests==2.32.3

